feat: adding groupName verification in group deletion

+++ backend/src/conversation/conversation.service.ts
@@ -269,7 +269,11 @@ export class ConversationService {
               picture: true,
             },
           }
-        }
+        },
+        orderBy: [
+          { owner: 'desc' },
+          { user: { name: 'asc' } }
+        ]
       });
 
       return {
@@ -342,7 +346,7 @@ export class ConversationService {
 
   }
 
-  async removeAll(loggedId:number, conversationId: number) {
+  async removeAll(loggedId: number, conversationId: number, groupName: string) {
     const userConversation = await this.prisma.userConversation.findUnique({
       where: {
         userId_conversationId: {
@@ -351,14 +355,27 @@ export class ConversationService {
         },
         owner: true
       },
+      include:{
+        conversation:{
+          select:{
+            name: true
+          }
+        }
+      }
     });
 
-    const conversation = await this.prisma.conversation.delete({
-      where: {
-        id: userConversation.conversationId
-      },
-    });
+    if (userConversation.conversation.name == groupName){
+      const conversation = await this.prisma.conversation.delete({
+        where: {
+          id: userConversation.conversationId
+        },
+      });
+      return {destroyMessage: "Grupo '" + conversation.name + "' deletado completamente."}
+    }
 
-    return {destroyMessage: "Grupo '" + conversation.name + "' deletado completamente."}
+    else{
+      return {wrongNameMessage: "O nome do grupo inserido está incorreto."}
+    }
+
   }
 }
